Refactor: move useBodyScrollLock to utils
