+++ src/Modal/index.ts
@@ -1,2 +1,1 @@
 export { Modal } from './Modal'
-export { useBodyScrollLock } from './useBodyScrollLock'
